Hacky Solution No 1 (log_noabort)
Add _noabort variants to
munit_logf_ex
munit_logf
munit_log

They do not stop the program, independent of log-level.

As a proof of concept, I added a _msgf variant to munit_assert,
that first logs the error with _noabort
and then the message with the default errorf.

This solution works, but I'd prefer some form of
indentention or grouping of user messages and error messages.

Edit: Changed to no longer refer to issue No 1.